[fix] 피드 조회 단건 -> 복수 선택 수정

+++ src/main/java/com/example/ForDay/domain/user/dto/response/GetUserFeedListResDto.java
@@ -11,7 +11,7 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class GetUserFeedListResDto {
-    private int totalFeedCount;
+    private Long totalFeedCount;
     private Long lastRecordId;
     private List<FeedDto> feedList;
     private boolean hasNext;
